split StudentTable into component
